Add cheap beam optimization
Only send beams to players that are facing the zone. Don't send them if
they are too far away as well. Cheap imitation of AddRecipientsByPVS.

This tries to stop sending beams to players if they wouldn't see them
anyways

If we only had access to the
IVEngineServer::Message_DetermineMulticastRecipients method..