Fix not triggering OnClientLeaveZone after team change
When a player changes his team while in a trigger_multiple, the trigger removes the player from its internal list of touching entities. If the player is spawned again right away inside the same trigger, the physics engine doesn't call |StartTouch| again, but just keeps calling |Touch|. The trigger_multiple doesn't add the player to its internal list again, because of the lack of the StartTouch callback. When the player leaves the zone now, the trigger ignores the |EndTouch| callback, because the player isn't registered in its internal list of touching entities.

To work around that, we call StartTouch manually if the |OnTrigger| output is called without |OnStartTouch| being called first. This adds the player to the internal list of the trigger correctly.